\name{getFontencAndLang}
\alias{getFontencAndLang}
\title{Determining the fontenc and language}
\description{getFontencAndLang returns the fontenc and language}
\usage{getFontencAndLang(mLangString=NA, mLangFile=NA)}
\arguments{	\item{mLangString}{ UTF-8 string}
	\item{mLangFile}{ UTF-8 file}}
\value{	character}
\author{Jose' Gama}
\keyword{programming}
