\name{is.African}
\alias{is.African}
\title{Determining if a UTF-8 string is within the African Unicode ranges}
\description{is.African returns TRUE if a UTF-8 string is within the African Unicode ranges}
\usage{is.African(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
