\name{is.Arabic}
\alias{is.Arabic}
\title{Determining if a UTF-8 string is within the Arabic Unicode ranges}
\description{is.Arabic returns TRUE if a UTF-8 string is within the Arabic Unicode ranges}
\usage{is.Arabic(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
