\name{is.Armenian}
\alias{is.Armenian}
\title{Determining if a UTF-8 string is within the Armenian Unicode ranges}
\description{is.Armenian returns TRUE if a UTF-8 string is within the Armenian Unicode ranges}
\usage{is.Armenian(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
