\name{is.ascii}
\alias{is.ascii}
\title{Determining if a UTF-8 character is ASCII}
\description{is.ascii returns TRUE if a UTF-8 character is ASCII
wrapper function for ascii from ctype.h}
\usage{is.ascii(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.ascii('a')
}
\author{Jose' Gama}
\keyword{programming}
