\name{is.CJK}
\alias{is.CJK}
\title{Determining if a UTF-8 string is within the CJK Unicode ranges}
\description{is.CJK returns TRUE if a UTF-8 string is within the CJK Unicode ranges}
\usage{is.CJK(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
