\name{is.Cyrillic}
\alias{is.Cyrillic}
\title{Determining if a UTF-8 string is within the Cyrillic Unicode ranges}
\description{is.Cyrillic returns TRUE if a UTF-8 string is within the Cyrillic Unicode ranges}
\usage{is.Cyrillic(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
