\name{is.Georgian}
\alias{is.Georgian}
\title{Determining if a UTF-8 string is within the Georgian Unicode ranges}
\description{is.Georgian returns TRUE if a UTF-8 string is within the Georgian Unicode ranges}
\usage{is.Georgian(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
