\name{is.Greek}
\alias{is.Greek}
\title{Determining if a UTF-8 string is within the Greek Unicode ranges}
\description{is.Greek returns TRUE if a UTF-8 string is within the Greek Unicode ranges}
\usage{is.Greek(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
