\name{is.greekmath}
\alias{is.greekmath}
\title{Determining if a UTF-8 string is within the greekmath Unicode ranges}
\description{is.greekmath returns TRUE if a UTF-8 string is within the greekmath Unicode ranges}
\usage{is.greekmath(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
