\name{is.Hangul}
\alias{is.Hangul}
\title{Determining if a UTF-8 string is within the Hangul Unicode ranges}
\description{is.Hangul returns TRUE if a UTF-8 string is within the Hangul Unicode ranges}
\usage{is.Hangul(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
