\name{is.Hebrew}
\alias{is.Hebrew}
\title{Determining if a UTF-8 string is within the Hebrew Unicode ranges}
\description{is.Hebrew returns TRUE if a UTF-8 string is within the Hebrew Unicode ranges}
\usage{is.Hebrew(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
