\name{is.Hiragana}
\alias{is.Hiragana}
\title{Determining if a UTF-8 string is within the Hiragana Unicode ranges}
\description{is.Hiragana returns TRUE if a UTF-8 string is within the Hiragana Unicode ranges}
\usage{is.Hiragana(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
