\name{is.isalnum}
\alias{is.isalnum}
\title{Determining if a UTF-8 character is alphanumeric}
\description{is.isalnum returns TRUE if a UTF-8 character is alphanumeric
wrapper function for isalnum from ctype.h}
\usage{is.isalnum(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
