\name{is.isalpha}
\alias{is.isalpha}
\title{Determining if a UTF-8 character is alphabetic}
\description{is.isalpha returns TRUE if a UTF-8 character is alphabetic
wrapper function for isalpha from ctype.h}
\usage{is.isalpha(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
