\name{is.isblank}
\alias{is.isblank}
\title{Determining if a UTF-8 character is blank}
\description{is.isblank returns TRUE if a UTF-8 character is blank
wrapper function for isblank from ctype.h}
\usage{is.isblank(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
