\name{is.iscntrl}
\alias{is.iscntrl}
\title{Determining if a UTF-8 character is a control code}
\description{is.iscntrl returns TRUE if a UTF-8 character is a control code
wrapper function for iscntrl from ctype.h}
\usage{is.iscntrl(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
