\name{is.isdigit}
\alias{is.isdigit}
\title{Determining if a UTF-8 character is alphanumeric}
\description{is.isdigit returns TRUE if a UTF-8 character is alphanumeric
wrapper function for isdigit from ctype.h}
\usage{is.isdigit(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
