\name{is.isgraph}
\alias{is.isgraph}
\title{Determining if a UTF-8 character has graphical representation}
\description{is.isgraph returns TRUE if a UTF-8 character has graphical representation
wrapper function for isgraph from ctype.h}
\usage{is.isgraph(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
