\name{is.islower}
\alias{is.islower}
\title{Determining if a UTF-8 character is lower case}
\description{is.islower returns TRUE if a UTF-8 character is lower case
wrapper function for islower from ctype.h}
\usage{is.islower(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
