\name{is.isprint}
\alias{is.isprint}
\title{Determining if a UTF-8 character is printable}
\description{is.isprint returns TRUE if a UTF-8 character is printable
wrapper function for isprint from ctype.h}
\usage{is.isprint(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
