\name{is.ispunct}
\alias{is.ispunct}
\title{Determining if a UTF-8 character is a punctuation character}
\description{is.ispunct returns TRUE if a UTF-8 character is a punctuation character
wrapper function for ispunct from ctype.h}
\usage{is.ispunct(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
