\name{is.isspace}
\alias{is.isspace}
\title{Determining if a UTF-8 character is space}
\description{is.isspace returns TRUE if a UTF-8 character is space
wrapper function for isspace from ctype.h}
\usage{is.isspace(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
