\name{is.isupper}
\alias{is.isupper}
\title{Determining if a UTF-8 character is upper case}
\description{is.isupper returns TRUE if a UTF-8 character is upper case
wrapper function for isupper from ctype.h}
\usage{is.isupper(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
