\name{is.isxdigit}
\alias{is.isxdigit}
\title{Determining if a UTF-8 character is an hexadecimal digit}
\description{is.isxdigit returns TRUE if a UTF-8 character is an hexadecimal digit
wrapper function for isxdigit from ctype.h}
\usage{is.isxdigit(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
