\name{is.JapaneseCJK}
\alias{is.JapaneseCJK}
\title{Determining if a UTF-8 string is within the JapaneseCJK Unicode ranges}
\description{is.JapaneseCJK returns TRUE if a UTF-8 string is within the JapaneseCJK Unicode ranges}
\usage{is.JapaneseCJK(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
