\name{is.JapaneseLatinAndHalfWidthKatakana}
\alias{is.JapaneseLatinAndHalfWidthKatakana}
\title{Determining if a UTF-8 string is within the JapaneseLatinAndHalfWidthKatakana Unicode ranges}
\description{is.JapaneseLatinAndHalfWidthKatakana returns TRUE if a UTF-8 string is within the JapaneseLatinAndHalfWidthKatakana Unicode ranges}
\usage{is.JapaneseLatinAndHalfWidthKatakana(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
