\name{is.JapanesePunctuation}
\alias{is.JapanesePunctuation}
\title{Determining if a UTF-8 string is within the JapanesePunctuation Unicode ranges}
\description{is.JapanesePunctuation returns TRUE if a UTF-8 string is within the JapanesePunctuation Unicode ranges}
\usage{is.JapanesePunctuation(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
