\name{is.Katakana}
\alias{is.Katakana}
\title{Determining if a UTF-8 string is within the Katakana Unicode ranges}
\description{is.Katakana returns TRUE if a UTF-8 string is within the Katakana Unicode ranges}
\usage{is.Katakana(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
