\name{is.LaTeX.reserved}
\alias{is.LaTeX.reserved}
\title{Determining if a UTF-8 string is within the LaTeX reserved Unicode ranges}
\description{is.LaTeX.reserved returns TRUE if a UTF-8 string is within the LaTeX reserved Unicode ranges}
\usage{is.LaTeX.reserved(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
