\name{is.Thai}
\alias{is.Thai}
\title{Determining if a UTF-8 string is within the Thai Unicode ranges}
\description{is.Thai returns TRUE if a UTF-8 string is within the Thai Unicode ranges}
\usage{is.Thai(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
