\name{is.Vietnamese}
\alias{is.Vietnamese}
\title{Determining if a UTF-8 string is within the Vietnamese Unicode ranges}
\description{is.Vietnamese returns TRUE if a UTF-8 string is within the Vietnamese Unicode ranges}
\usage{is.Vietnamese(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\keyword{programming}
