\name{joinListOfTags}
\alias{joinListOfTags}
\title{ Create a file with tag data in three colums from a list of files}
\description{joinListOfTags creates a file with 3 columns (UTF8 code points, UTF8 character and LaTeX tag) from a list of files }
\usage{joinListOfTags(files3colsUTFlatex, fileCombined)}
\arguments{	\item{files3colsUTFlatex}{ UTF-8 file with tags in three columns}
	\item{fileCombined}{ vector with a list of filenames}}
\value{	character}
\author{Jose' Gama}
\keyword{programming}
