\name{langID}
\alias{langID}
\title{Determining the language of a UTF8 point}
\description{langID returns a character indicating the language of a UTF8 point
African		'F'
Thai		'T'
Vietnamese	'V'
Cyrillic	'R'
Greek		'G'
Hebrew		'H'
Arabic		'A'
Armenian	'N'
Georgian	'E'
Latin		'L'
}
\usage{langID(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	character}
\author{Jose' Gama}
\keyword{programming}
