\name{is.alnum}
\alias{is.alnum}
\title{Determining if a UTF-8 character is alphanumeric}
\description{is.alnum returns TRUE if a UTF-8 character is alphanumeric
wrapper function for isalnum from ctype.h}
\usage{is.alnum(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\author{Jose' Gama}
\examples{
is.alnum('7')
}
\keyword{programming}
