\name{is.alpha}
\alias{is.alpha}
\title{Determining if a UTF-8 character is alphabetic}
\description{is.alpha returns TRUE if a UTF-8 character is alphabetic
wrapper function for isalpha from ctype.h}
\usage{is.alpha(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.alpha('a')
}
\author{Jose' Gama}
\keyword{programming}
