\name{is.blank}
\alias{is.blank}
\title{Determining if a UTF-8 character is blank}
\description{is.blank returns TRUE if a UTF-8 character is blank
wrapper function for isblank from ctype.h}
\usage{is.blank(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.blank('a')
}
\author{Jose' Gama}
\keyword{programming}
