\name{is.digit}
\alias{is.digit}
\title{Determining if a UTF-8 character is alphanumeric}
\description{is.digit returns TRUE if a UTF-8 character is alphanumeric
wrapper function for isdigit from ctype.h}
\usage{is.digit(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.digit('7')
}
\author{Jose' Gama}
\keyword{programming}
