\name{is.graph}
\alias{is.graph}
\title{Determining if a UTF-8 character has graphical representation}
\description{is.graph returns TRUE if a UTF-8 character has graphical representation
wrapper function for isgraph from ctype.h}
\usage{is.graph(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.graph('a')
}
\author{Jose' Gama}
\keyword{programming}
