\name{is.lower}
\alias{is.lower}
\title{Determining if a UTF-8 character is lower case}
\description{is.lower returns TRUE if a UTF-8 character is lower case
wrapper function for islower from ctype.h}
\usage{is.lower(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.lower('a')
}
\author{Jose' Gama}
\keyword{programming}
