\name{is.print}
\alias{is.print}
\title{Determining if a UTF-8 character is printable}
\description{is.print returns TRUE if a UTF-8 character is printable
wrapper function for isprint from ctype.h}
\usage{is.print(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.print('a')
}
\author{Jose' Gama}
\keyword{programming}
