\name{is.punct}
\alias{is.punct}
\title{Determining if a UTF-8 character is a punctuation character}
\description{is.punct returns TRUE if a UTF-8 character is a punctuation character
wrapper function for ispunct from ctype.h}
\usage{is.punct(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.punct('!')
}
\author{Jose' Gama}
\keyword{programming}
