\name{is.space}
\alias{is.space}
\title{Determining if a UTF-8 character is space}
\description{is.space returns TRUE if a UTF-8 character is space
wrapper function for isspace from ctype.h}
\usage{is.space(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.space(' ')
}
\author{Jose' Gama}
\keyword{programming}
