\name{is.upper}
\alias{is.upper}
\title{Determining if a UTF-8 character is upper case}
\description{is.upper returns TRUE if a UTF-8 character is upper case
wrapper function for isupper from ctype.h}
\usage{is.upper(utf8char)}
\arguments{	\item{utf8char}{ UTF-8 string}}
\value{	logic}
\examples{
is.upper('A')
}
\author{Jose' Gama}
\keyword{programming}
